Add fallback for suphp to the PHP config hook
Case ZC-891:  We were assuming that suphp would always be available, by
using it for our default handler for a newly-installed, unconfigured PHP
version.  When mod_suphp is *not* installed, however, it will cause
failures.  We will now check whether mod_suphp is installed before trying
to set it as the default, and if not we'll fall back to cgi.  Our PHP
packages install in such a way that the cgi handler can be guaranteed to
always be there.