Verify that the requested system default is installed
Case EA-3608:  We were not checking to verify whether the system default
PHP version that we had previously set was still installed, which could
cause errors.  We now pull the system default, compare it and the contents
we retrieve from the YAML file against the list of actually installed PHP
versions, and set a legal version.

Also refactored the construction of the installed PHP versions list, and
handling of any failures due to none being installed.