Added missing global "/" directory
Case ZC-914: Users unable to apply RewriteEngine rules to
htaccess file because global "/" AllowOverride option was
missing.

NOTE: It would be better from a security stand-point to
deny all access to "/", so that everything inherits from that.
After denying access, this would require us to explicity open
holes to other parts of the file system (e.g. a whitelist
system, as opposed to blanketly allow everything through).

However I'm unsure of the implications this would cause, and
will resort of maintaining the EA3 status quo.