Updated proxy subdomains for cpanel & whm
Case ZC-920: This principally adds the missing proxy subdomains to
the apache configuration file.  However, it also moves the
virtual hosts blocks to the end.  This ensures that all cPanel & WHM
configurations settings are at the top, and all user domains are
at the bottom.

This also fixes CPANEL-1029 by updating the RewriteRule rule for
proxy subdomains.  "\$1" was being used instead of "$1".  This lead
to an improper URL for cPanel.  Thanks Julian for finding this.

NOTE: As of this writing, you need to manually install the
following package to make it proxy subdomains work:
 ea-apache24-mod_proxy_http