Rollback new modsec conf locations until WHM can support it.
Case EA-4789:

Merge upstreasm of Case 4673.

The new RPM will migrate modsec2 conf files to conf.d/modsec
but WHM might not be updated enough to write to them in the
new location. This post-RPM installation trigger will check
the version of WHM and if it is not new enough, will un-migrate
the conf.d/modsec files (and changes to conf.d/modsec2.conf).

When WHM eventually becomes new enough to configure the files in
the new locations, this script will perform no actions, allowing the
next RPM installation of mod_modsecurity2 to move the files without
reverting.