Fix bad warning about apache handler removal when installing.
Case EA-4914:

Previously any apache handler change, including adding a new handler,
reported a warning that a configured Apache handler was removed.

Now only handler changes trigger a warning that a configured Apache
handler was removed (since this is in a yum transaction, the only
trigger for this should be handler removals).