Add support for SymlinkProtect in httpd.conf templates.
Case EA-5487:

This commit adds support for including SymlinkProtect and
SymlinkProtectRoot in the httpd.conf templates, such that if
symlink protection is on, the two directives are present.  If it is
off, the two directives are absent.

Manual configuration supports more configuraiton combinations than
these, but these combinations are sufficent for UI support.