Relocate run directory to avoid cross-filesystem access.
Case EA-5533:

Previously, httpd would try to access the PID file in /etc/apache2/run
which mapped to /var/run/apache2 which on CentOS 7 was further mapped
to /run/apache2.

After extensive testing, it was determined that apr is not always able
to read /run/apache2/httpd.pid by opening /etc/apache2/run/httpd.pid,
and this failure to detect the file led the second (new) httpd process
to think there was only one httpd process, leading to two longstanding
root httpd processes.

This failure to read is not 100% stable, but it is documented in the
APR as a known issue.  Therefore, we have relocated the paths by which
http.pid is opened to not cross filesystem boundaries.  In other words
CentOS 6 systems open /var/run/apache2/httpd.pid directly and CentOS 7
systems open /run/apache2/httpd.pid directly.