Don’t bring in SSLCertificateKeyFile if none is given.
Case EA-6682: cPanel & WHM 68 anticipates updating the indexing and
storage of SSL certificate chains and keys to accommodate the radical
uptick in SSL usage since introduction of AutoSSL in v58.

This will involve combining the key with the full certificate chain
into a single file, to be given in the SSLCertificateFile directive.
The SSLCertificateKeyFile directive will not be needed.

From the templates’ perspective, the only difference will be that the
key file is no longer given; thus, we need the templates to know not to
put in that directive if the key file isn’t given.

(Cloned from COBRA-5343)