Add proxy configuration for ActiveSync.
Case CPANEL-36164: For ActiveSync support in cPanel & WHM, the
service is accepting connections on port 2091, but we need to provide
a mechanism similar to proxy subdomains for clients that are unable
to use that port (for example, due to firewalls blocking traffic
on unusual ports). For ActiveSync, it makes more sense to offer a
"proxy path" in the main domain than it does to require use of a
separate subdomain because this fits better into Android devices'
setup workflow, where the server hostname is derived from the domain
of an email address.

This doesn't fit the pattern of the existing proxy subdomains,
so it can't just be another entry in the vhost.proxy_subdomains
data structure.

Team merge task: BWG-2217