Avoid redirection loops when handling requests from reverse proxies.
Case CPANEL-36901: When “Force HTTPS Redirect” is enabled in cPanel, and
a user access the non-https URL for that domain through a reverse proxy,
it was possible for the user to get put into an infinite redirection
loop.

This change addresses that by adding a check for the
X-Forwarded-Proto header sent by those proxies, ensuring that it doesn't
redirect when accessing over the https protcol.

Changelog: Fix redirection loop when forcing HTTPS redirects.